+++ trackAndTrace/frontend/src/components/AdminCreateItem.tsx
@@ -16,6 +16,7 @@ import * as constants from '../constants';
 import { TxHashLink } from './CCDScanLinks';
 import { createItem } from '../track_and_trace_contract';
 import * as TrackAndTraceContract from '../../generated/module_track_and_trace';
+import { FromTokenIdU64 } from '../utils';
 
 interface Props {
     connection: WalletConnection | undefined;
@@ -24,7 +25,7 @@ interface Props {
 }
 
 interface PartialItemCreatedEvent {
-    item_id: number | bigint;
+    item_id: string;
 }
 
 export function AdminCreateItem(props: Props) {
@@ -65,7 +66,12 @@ export function AdminCreateItem(props: Props) {
                         const parsedEvent = TrackAndTraceContract.parseEvent(eventList.events[0]);
                         const itemCreatedEvent = parsedEvent.content as unknown as PartialItemCreatedEvent;
 
-                        setNewItemId(itemCreatedEvent.item_id);
+                        // The `item_id` is of type `TokenIdU64` in the smart contract and logged in the event as
+                        // a little-endian hex string.
+                        // E.g. the `TokenIdU64` representation of `1` is the hex string `0100000000000000`.
+                        // This function converts the `TokenIdU64` representation into a bigint type here.
+                        const itemId: bigint = FromTokenIdU64(itemCreatedEvent.item_id);
+                        setNewItemId(itemId);
                     } else {
                         setItemIdError('Tansaction failed and event decoding failed.');
                     }