+++ trackAndTrace/frontend/src/components/ChangeItemStatus.tsx
@@ -13,6 +13,7 @@ import { TxHashLink } from './CCDScanLinks';
 import * as constants from '.././constants';
 import { nonceOf } from '../track_and_trace_contract';
 import * as TrackAndTraceContract from '../../generated/module_track_and_trace'; // Code generated from a smart contract module. The naming convention of the generated file is `moduleName_smartContractName`.
+import { ToTokenIdU64 } from '../utils';
 
 interface Props {
     connection: WalletConnection | undefined;
@@ -29,7 +30,7 @@ const NEW_STATUS_OPTIONS = [
 
 function generateMessage(
     newStatus: 'Produced' | 'InTransit' | 'InStore' | 'Sold' | undefined,
-    itemID: bigint,
+    itemID: number | bigint,
     expiryTimeSignature: Date,
     nonce: number | bigint,
 ) {
@@ -38,12 +39,16 @@ function generateMessage(
             throw Error(`'newStatus' input field is undefined`);
         }
 
+        // The `item_id` is of type `TokenIdU64` in the smart contract which is represented as a little-endian hex string.
+        // E.g. the `TokenIdU64` representation of `1` is the hex string `0100000000000000`.
+        const tokenIdU64 = ToTokenIdU64(itemID);
+
         // Create ChangeItemStatus parameter
         const changeItemStatusParameter: TrackAndTraceContract.ChangeItemStatusParameter = {
             additional_data: {
                 bytes: [],
             },
-            item_id: Number(itemID),
+            item_id: tokenIdU64,
             new_status: {
                 type: newStatus,
             },
@@ -71,7 +76,7 @@ export function ChangeItemStatus(props: Props) {
     const { connection, accountAddress, activeConnectorError } = props;
 
     interface FormType {
-        itemID: bigint | undefined;
+        itemID: number | bigint | undefined;
         newStatus: 'Produced' | 'InTransit' | 'InStore' | 'Sold' | undefined;
     }
     const { control, register, formState, handleSubmit } = useForm<FormType>({ mode: 'all' });