+++ trackAndTrace/frontend/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased changes
 
+## 1.2.0
+
+-   Display the `initial_status` field from the create event.
+-   Change to match the `TokenIdU64` type for the `item_id` field in the create event and update event.
+
 ## 1.1.0
 
 -   Add transition edges management by admin to front end.